move mirror to a standalone package
