+++ routers/init.go
@@ -21,6 +21,7 @@ import (
 	"code.gitea.io/gitea/modules/setting"
 	"code.gitea.io/gitea/modules/ssh"
 	"code.gitea.io/gitea/services/mailer"
+	"code.gitea.io/gitea/services/mirror_service"
 
 	"gitea.com/macaron/macaron"
 )
@@ -98,7 +99,7 @@ func GlobalInit() {
 			log.Fatal("Failed to initialize issue indexer: %v", err)
 		}
 		models.InitRepoIndexer()
-		models.InitSyncMirrors()
+		mirror_service.InitSyncMirrors()
 		models.InitDeliverHooks()
 		models.InitTestPullRequests()
 	}
